2:41 restrictedpage

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -14,6 +16,10 @@ from .forms import RoomForm
 #     {'id':3, 'name':'Lets learn english!'}
 # ]
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -29,12 +35,15 @@ def loginPage(request):
             messages.error(request, 'Username or password does not exist')
 
 
-    context = {}
+    context = {'page'}
     return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+def registerUser(request):
+    page ='register'
+    return render(request,'base/login_register.html')
 
 
 def home(request):
@@ -55,6 +64,8 @@ def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(request,'base/room.html',context)
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -66,9 +77,14 @@ def createRoom(request):
     context ={'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
+
     context={'form': form}
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -77,9 +93,11 @@ def updateRoom(request, pk):
             return redirect('home')
     return render(request,'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here.")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
